Fused pieces ShortElbow and ShortElbowAngle. Moved old ShortElbow -> ShortElbow90 (to keep this code). Renamed ShortElbowAngle -> ShortElbow (for simplicity). Updated corresponding references in tests and ui.
