+++ nfluid/elements/short_elbow_angle_auto.py
@@ -79,7 +79,7 @@ class ShortElbowAngleAuto(ChannelElement2G):
 
     def create_shape_child(self):
         print 'create_shape ShortElbowAngleAuto'
-        return CreateShape('short_elbow_angle', self.CenterPos,
+        return CreateShape('short_elbow', self.CenterPos,
                            self.RotationOperator,
                            self.get_r(), self.angle,
                            self.get_pos_head(), self.get_pos_tail(),
