+++ nfluid/tests/TestShapeShortElbowAngle.py
@@ -4,7 +4,7 @@ import os
 import sys
 from nfluid.core.channel_assembly import ChannelAssembly, create_channel
 from nfluid.elements.circle_coupling import CircleCoupling
-from nfluid.elements.short_elbow_angle import ShortElbowAngle
+from nfluid.elements.short_elbow import ShortElbow
 from nfluid.tests.TestBase import MakeTest1
 from nfluid.util.vector import Vector
 import math
@@ -28,16 +28,16 @@ if sys.argv[1] == '1':
     print 'Test_1'
     create_channel(CircleCoupling(R=10, L=20, PosH=Vector(0, 0, 0),
                                   Normal=Vector(0, 0, 1))
-                   .link(ShortElbowAngle(Angle=45, NormalT=Vector(1, 0, 1)))
+                   .link(ShortElbow(Angle=45, NormalT=Vector(1, 0, 1)))
                    .link(CircleCoupling(L=50))
-                   .link(ShortElbowAngle(Angle=45, NormalT=Vector(0, 0, 1)))
+                   .link(ShortElbow(Angle=45, NormalT=Vector(0, 0, 1)))
                    .link(CircleCoupling(L=30)))
 
 elif sys.argv[1] == '2':
     print 'Test_2'
     create_channel(CircleCoupling(R=20, L=40, PosH=Vector(0, 0, 0),
                                   Normal=Vector(0, 0, 1))
-                   .link(ShortElbowAngle(Angle=45, NormalT=Vector(0, 1, 1)))
+                   .link(ShortElbow(Angle=45, NormalT=Vector(0, 1, 1)))
                    .link(CircleCoupling(L=80)))
 
 elif sys.argv[1] == '3':
@@ -45,7 +45,7 @@ elif sys.argv[1] == '3':
     sqrt2 = math.sqrt(2)
     create_channel(CircleCoupling(R=5, L=30, PosH=Vector(0, 0, 0),
                                   Normal=Vector(0, 0, 1))
-                   .link(ShortElbowAngle(Angle=135,
+                   .link(ShortElbow(Angle=135,
                                          NormalT=Vector(1, 1, -sqrt2)))
                    .link(CircleCoupling(L=80)))
 
@@ -53,9 +53,9 @@ elif sys.argv[1] == '4':
     print 'Test_4'
     create_channel(CircleCoupling(R=10, L=20, PosH=Vector(0, 0, 0),
                                   Normal=Vector(0, 0, 1))
-                   .link(ShortElbowAngle(Angle=90, NormalT=Vector(1, 0, 0)))
+                   .link(ShortElbow(Angle=90, NormalT=Vector(1, 0, 0)))
                    .link(CircleCoupling(L=50))
-                   .link(ShortElbowAngle(Angle=90, NormalT=Vector(0, 0, 1)))
+                   .link(ShortElbow(Angle=90, NormalT=Vector(0, 0, 1)))
                    .link(CircleCoupling(L=30)))
 
 elif sys.argv[1] == '0':
