+++ nfluid/ui/elements/tabpieceswidget.py
@@ -15,7 +15,8 @@ _pieces = [(strings.coupling,
             ),
            (strings.short_elbow,
             [(strings.head_radius, 0),
-             (strings.tail_normal, (0, 0, 0))]
+             (strings.tail_normal, (0, 0, 0)),
+             (strings.angle, 0)]
             ),
            (strings.long_elbow,
             [(strings.curvature_radius, 0),
